+++ pyproject.toml
@@ -23,6 +23,7 @@ nemosis = "^3.4.0"
 mypy = "^1.9.0"
 jupyter = "^1.0.0"
 pandas-stubs = "^2.2.1.240316"
+pytest = "^8.2.2"
 
 
 [build-system]
