+++ tm_solarshift/models/resistive_single.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 import numpy as np
 import pandas as pd
-from typing import TYPE_CHECKING, TypeAlias
 
 from tm_solarshift.constants import DIRECTORY
 from tm_solarshift.utils.units import (Variable, Water)
@@ -15,44 +14,45 @@ from tm_solarshift.models.hw_tank import (
 
 FILES_MODEL_SPECS = DIRECTORY.FILES_MODEL_SPECS
 
-class HeatPump(HWTank):
+class ResistiveSingle(HWTank):
     def __init__(self):
 
         # description
-        self.name = "Heat Pump, external heat exchanger with thermostat."
-        self.label = "heat_pump"
+        self.name = "Conventional resistive immersive heater (single unit)."
+        self.label = "resistive"
         self.model = "-"
         self.cost = Variable(np.nan, "AUD")
 
-
-        # heater
-        self.nom_power_th = Variable(5240.0, "W")
-        self.nom_power_el = Variable(870.0, "W")
-        self.eta = Variable(6.02, "-")
-        self.nom_tamb = Variable(32.6, "degC")
-        self.nom_tw = Variable(21.1, "degC")
+        #Loading all default values
+        # heater data
+        self.nom_power = Variable(3600.0, "W")
+        self.eta = Variable(1.0, "-")
 
         super().__init__()
-        # tank
-    #     self.vol = Variable(0.315,"m3")
-    #     self.height = Variable(1.45, "m")  # It says 1.640 in specs, but it is external height, not internal
-    #     self.height_inlet = Variable(0.113, "m")
-    #     self.height_outlet = Variable(1.317, "m")
-    #     self.height_heater = Variable(0.103, "m")
-    #     self.height_thermostat = Variable(0.103, "m")
-    #     self.U = Variable(0.9, "W/m2-K")
 
-    #     #numerical simulation
-    #     self.fluid = Water()
-    #     self.nodes = 10     # Tank nodes. DO NOT CHANGE, unless TRNSYS layout is changed too!
-    #     self.temps_ini = 3  # [-] Initial temperature of the tank. Check trnsys.editing_dck_tank() for options
+        # # tank geometry and losses
+        # self.vol = Variable(0.315,"m3")
+        # self.height = Variable(1.45, "m")  # It says 1.640 in specs, but it is external height, not internal
+        # self.height_inlet = Variable(0.113, "m")
+        # self.height_outlet = Variable(1.317, "m")
+        # self.height_heater = Variable(0.103, "m")
+        # self.height_thermostat = Variable(0.103, "m")
+        # self.U = Variable(0.9, "W/m2-K")
 
-    #     # control
-    #     self.temp_max = Variable(65.0, "degC")  #Maximum temperature in the tank
-    #     self.temp_deadband = Variable(10.0, "degC") # Dead band for max temp control
-    #     self.temp_min = Variable(45.0, "degC")  # Minimum temperature in the tank
-    #     self.temp_consump = Variable(45.0, "degC") #Consumption temperature
+        # #numerical simulation
+        # self.fluid = Water()
+        # self.nodes = 10     # Tank nodes. DO NOT CHANGE, unless TRNSYS layout is changed too!
+        # self.temps_ini = 3  # [-] Initial temperature of the tank. Check trnsys.editing_dck_tank() for options
 
+        # # control
+        # self.temp_max = Variable(65.0, "degC")  #Maximum temperature in the tank
+        # self.temp_deadband = Variable(10.0, "degC") # Dead band for max temp control
+        # self.temp_min = Variable(45.0, "degC")  # Minimum temperature in the tank
+        # self.temp_consump = Variable(45.0, "degC") #Consumption temperature
+
+    def __eq__(self, other) : 
+            return self.__dict__ == other.__dict__
+    
     # @property
     # def tank_thermal_cap(self) -> Variable:
     #     return tank_thermal_capacity(self)
@@ -66,12 +66,11 @@ class HeatPump(HWTank):
     # def tank_temp_high_control(self) -> Variable:
     #     return tank_temp_high_control(self)
 
-
     @classmethod
     def from_model_file(
         cls,
-        file_path: str = FILES_MODEL_SPECS["heat_pump"],
-        model: str = "",
+        file_path: str = FILES_MODEL_SPECS["resistive"],
+        model:str = "",
         ):
         df = pd.read_csv(file_path, index_col=0)
         specs = pd.Series(df.loc[model])
@@ -84,8 +83,8 @@ class HeatPump(HWTank):
             except:
                 pass          
             setattr(output, lbl, Variable(value, unit) )
+
         return output
-    
     
     def run_thermal_model(
             self,
@@ -96,3 +95,7 @@ class HeatPump(HWTank):
         trnsys_dewh = TrnsysDEWH(DEWH=self, ts=ts)
         df_tm = trnsys_dewh.run_simulation(verbose=verbose)
         return df_tm
+    
+if __name__ == "__main__":
+    DEWH = ResistiveSingle.from_model_file(model="491315")
+    print(DEWH.vol)
