+++ tm_solarshift/models/trnsys.py
@@ -14,7 +14,8 @@ from tm_solarshift.utils.units import (Variable, conversion_factor as CF)
 if TYPE_CHECKING:
     from tm_solarshift.general import Simulation
     from tm_solarshift.utils.location import Location
-    from tm_solarshift.models.dewh import ( ResistiveSingle, HeatPump)
+    from tm_solarshift.models.resistive_single import ResistiveSingle
+    from tm_solarshift.models.heat_pump import HeatPump
     from tm_solarshift.models.gas_heater import GasHeaterStorage
     from tm_solarshift.models.solar_thermal import SolarThermalElecAuxiliary
     Heater: TypeAlias = ResistiveSingle | HeatPump | GasHeaterStorage | SolarThermalElecAuxiliary
